pyproject.toml
[project]
name = "ai-sales-trainer"
version = "0.1.0"
description = "AI Sales Training Chatbot - Multi-Agent System"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "litellm>=1.60.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.21.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.0",
    "sse-starlette>=2.2.0",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-jose[cryptography]>=3.5.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
